regarding issue about ints
see: https://github.com/IL2232-HT2022-HIL/SOFTWARE_DEV/issues/62

+++ Board-code/HiL_Board/HiL_code/Inc/HiL_gateway.h
@@ -14,7 +14,7 @@
 #include "usbd_cdc_if.h" //for transmission over USB
 
 
-void   HiL_gateway_transmit_message(int value1, int value2);
+void HiL_gateway_transmit_message(uint8_t value1, uint8_t value2);
 
 
 #endif /* GATEWAY__GATEWAY_H_ */
