+++ Board-code/HiL_Board/HiL_code/Src/HiL_gateway.c
@@ -8,13 +8,11 @@
 
 #include "HiL_gateway.h"
 
-
-void HiL_gateway_transmit_message(int value1, int value2)
+void HiL_gateway_transmit_message(uint8_t value1, uint8_t value2)
 {
 	uint8_t values[2];
 	values[0] = value1;
 	values[1] = value2;
-
 
 	CDC_Transmit_FS( /*(uint8_t *)*/ values, sizeof(values));		// Transmit what's been recieved in our msg queue
 }
